Add Sorted() method to addrdata
AddrDatas is a data structure that associates arbitrary data to
non-overlapping address ranges. Iterating through this structure
happens in the order of starting addresses. This change adds Sorted()
function to the data structure that enables sorting the address ranges
based on the data associated to them. For instance, if data is a
timestamp of last access, Sorted() helps sorting ranges according to
time since accessed.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>