trackers: fix softdirty and idlepage sudden tracking stops
If a process got completely swapped out, trackers got only an empty
list of address ranges and treated this as a signal of the whole
process having disappeared. As a result, they effectively stopped
tracking the process. Change the behavior so that only an error in
reading address ranges means that the processes has disappeared, and
in that case, remove also tracking counters of that process. That is,
GetCounters() will return data only on processes that are thought to
be running.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>