e2e: always clean up meme cgroup if it was in use
memtierd-meme-stop does not need to sleep unless it is removing the
meme cgroup. The cgroup should be removed regardless of MEME_PID
environment variable, but if removing it fails, it is not an error of
stopping meme. In that case there might be something else running in
the cgroup, and we will let them run. Memes got killed anyway.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>