trackers: add new "finder" tracker
The finder tracker is the most lightweight tracker that can be used to
only report process memory ranges without any page access
information. Instead of all memory ranges, the tracker can be
configured to report pages that have certain pagemap bits set or
cleared, too. For instance pages that are present in memory, pages
that are swapped, or pages that are neither present nor swapped.