e2e: more robust heat/idlepage test
The idlepage test failes if vm does not have THP enabled, or if meme
has been built with make DEBUG=1. This patch fixes the first issue by
making sure that THP is enabled, and the second issue by relaxing the
amount of cold memory that is expected to be moved.