e2e: add python3 starting and commands
memtierd-python3-start() enables starting multiple Python 3 processes,
each on any cgroup that can be created on-the-fly.

memtierd-python3-command() sends a command to selected Python 3
process. It prints a prompt that identifies the Python 3 process,
command that will be executed, and response from the process. Commands
and outputs are logged for debugging.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>