memtierd: implement new policy: avoid-oom
The avoid-oom policy scans cpuset.mems of tracked cgroups and free
memory available on each NUMA node in the system. When the policy
detects a risk of OOM due to total memory in all nodes of some
cpuset.mems running low, it looks for options to balance data in
memory. For instance, if a process has memory on nodes under pressure,
but it is allowed to use memory on nodes with no pressure, too, then
the policy moves its memory from high pressure to no pressure nodes,
thereby freeing memory on high-pressure nodes and avoiding OOM
killing. Processed whose memory gets moves are selected based on
amount of memory and OOM score.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>